% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir.R
\name{create_dir}
\alias{create_dir}
\title{Create Directory and Return Path}
\usage{
create_dir(path, recursive = TRUE, absolute = TRUE)
}
\arguments{
\item{path}{Character. Directory path to create.}

\item{recursive}{Logical. Whether to create parent directories if they don't exist. Default is TRUE.}

\item{absolute}{Logical. Whether to return absolute path. Default is TRUE.}
}
\value{
A character string of the created directory path.
}
\description{
Create Directory and Return Path
}
\examples{
create_dir("output")
create_dir("output/results", absolute = FALSE)
}
